// src/types.ts
export interface WorkspaceDimensions {
  width: number
  depth: number
}

export interface FootprintSize {
  w: number
  d: number
  h: number
}

export interface SafeZoneBounds {
  minX: number
  maxX: number
  minZ: number
  maxZ: number
}

export interface Position3D {
  x: number
  y: number
  z: number
}